+++ lib/object_forge/forge.rb
@@ -2,6 +2,7 @@
 
 require_relative "crucible"
 require_relative "forge_dsl"
+require_relative "molds"
 
 module ObjectForge
   # Object instantitation forge.
@@ -19,14 +20,22 @@ module ObjectForge
     # @!attribute [r] traits
     #   Attributes belonging to traits.
     #   @return [Hash{Symbol => Hash{Symbol => Any}}]
-    Parameters = Struct.new(:attributes, :traits, keyword_init: true)
+    #
+    # @!attribute [r] mold
+    #   An object that knows how to build the instance.
+    #   Must have a +call+ method that takes a class and a hash of attributes.
+    #   @since 0.2.0
+    #   @return [#call, nil]
+    Parameters = Struct.new(:attributes, :traits, :mold, keyword_init: true)
+
+    MOLD_MOLD = Molds::MoldMold.new.freeze
 
     # Define (and create) a forge using DSL.
     #
     # @see ForgeDSL
     # @thread_safety Thread-safe if DSL definition is thread-safe.
     #
-    # @param forged [Class] class to forge
+    # @param forged [Class, Any] class or object to forge
     # @param name [Symbol, nil] forge name
     # @yieldparam f [ForgeDSL]
     # @yieldreturn [void]
@@ -38,19 +47,20 @@ module ObjectForge
     # @return [Symbol, nil] forge name
     attr_reader :name
 
-    # @return [Class] class to forge
+    # @return [Class, Any] class or object to forge
     attr_reader :forged
 
     # @return [Parameters, ForgeDSL] forge parameters
     attr_reader :parameters
 
-    # @param forged [Class] class to forge
+    # @param forged [Class, Any] class or object to forge
     # @param parameters [Parameters, ForgeDSL] forge parameters
     # @param name [Symbol, nil] forge name
     def initialize(forged, parameters, name: nil)
       @name = name
       @forged = forged
       @parameters = parameters
+      @mold = parameters.mold || MOLD_MOLD.call(forged: forged)
     end
 
     # Forge a new instance.
@@ -67,7 +77,7 @@ module ObjectForge
     # If a block is given, forged instance is yielded to it after being built.
     #
     # @thread_safety Forging is thread-safe if {#parameters},
-    #    +traits+ and +overrides+ are thread-safe.
+    #   +traits+ and +overrides+ are thread-safe.
     #
     # @param traits [Array<Symbol>] traits to apply
     # @param overrides [Hash{Symbol => Any}] attribute overrides
@@ -76,7 +86,7 @@ module ObjectForge
     # @return [Any] built instance
     def forge(*traits, **overrides)
       resolved_attributes = resolve_attributes(traits, overrides)
-      instance = build_instance(resolved_attributes)
+      instance = @mold.call(forged: @forged, attributes: resolved_attributes)
       yield instance if block_given?
       instance
     end
@@ -89,10 +99,6 @@ module ObjectForge
     def resolve_attributes(traits, overrides)
       attributes = @parameters.attributes.merge(*@parameters.traits.values_at(*traits), overrides)
       Crucible.new(attributes).resolve!
-    end
-
-    def build_instance(attributes)
-      forged.new(attributes)
     end
   end
 end